utils.parse: refactored + created evaluate
